database/requirements.txt
Django==2.0.6
django_filters==0.2.1
djangorestframework==3.8.2
